Cleaned up the Transform class and removed some warnings by changing some variables' data types and adding static_cast where needed

+++ Sail/src/API/DX12/shader/DX12InputLayout.cpp
@@ -37,7 +37,7 @@ void DX12InputLayout::pushVec4(InputType inputType, LPCSTR semanticName, UINT se
 void DX12InputLayout::create(void* vertexShaderBlob) {
 	
 	m_inputLayoutDesc.pInputElementDescs = m_inputElementDescs.data();
-	m_inputLayoutDesc.NumElements = m_inputElementDescs.size();
+	m_inputLayoutDesc.NumElements = static_cast<UINT>(m_inputElementDescs.size());
 
 }
